# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Cafe POS API
  version: 1.0.0
  description: REST contract used by the Cafe POS React frontend.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/self-order/menu/{token}:
    get:
      tags: [Self order]
      security: []
      parameters:
        - {$ref: '#/components/parameters/Token'}
      responses:
        '200': {description: Public table menu}
  /api/self-order/order/{token}:
    post:
      tags: [Self order]
      security: []
      parameters:
        - {$ref: '#/components/parameters/Token'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SelfOrderRequest'}
      responses:
        '200': {description: Self order placed}
  /api/self-order/order/{orderId}/status:
    get:
      tags: [Self order]
      security: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        '200': {description: Order and kitchen status}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Token:
      name: token
      in: path
      required: true
      schema: {type: string}
...
  schemas:
...
    OrderLineRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId: {type: integer, format: int64}
        quantity: {type: number, minimum: 0.001}
...
    SelfOrderRequest:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/OrderLineRequest'}
...